cloud_jobs/news_scraper/requirements.txt
# Core
flask>=2.3.0
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
lxml>=4.9.0
lxml_html_clean>=0.1.1

# Scraper specific
newspaper3k>=0.2.8
gnews>=0.3.6
googlenewsdecoder>=0.1.1
trafilatura>=1.6.0

# Sentiment model
transformers>=4.30.0
google-cloud-storage>=2.10.0

# Optional / Database
python-dotenv>=1.0.0
oracledb>=1.4.0
